+++ lib/liquid_tabbar_minimize.dart
@@ -1,5 +1,3 @@
 export 'src/liquid_bottom_navigation_bar.dart';
-export 'src/liquid_tab_item.dart';
-export 'src/liquid_tabbar.dart';
 export 'src/liquid_route_observer.dart';
 export 'liquid_tabbar_minimize_android.dart';